[Refact] 새로운 게시물 페이지 리팩토링
index에 너무 많은 로직이 존재하고 있어 util성 파일과 library성 파일과
custom hooks를 만들어서 분리하는 방향으로 리팩토링.
