+++ web/src/containers/NewPostPage/NewPostWrapper.js
@@ -15,6 +15,16 @@ const NewPostWrapper = styled.div`
     color: black;
     margin-left: 10px;
   }
+
+  .crop-container {
+    position: relative;
+    width: 650px;
+    height: 650px;
+  }
+
+  .controls {
+    width: 20%;
+  }
 `;
 
 export default NewPostWrapper;
