+++ web/src/containers/NewPostPage/index.js
@@ -1,12 +1,15 @@
-import React, { useReducer, useCallback, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { Redirect } from 'react-router-dom';
 import Slider from '@material-ui/core/Slider';
 import Cropper from 'react-easy-crop';
 import { ToastContainer, toast, Slide } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import getCroppedImg from '../../utils/cropImage';
+import { isFileTypeImage } from '../../utils/fileUtils';
+import makeNewImageFile from './lib/makeNewImageFile';
+import useImage from './hooks/useImage';
 import Loading from '../../components/Loading';
 import Button from '../../components/Button';
+import Error from '../../components/Error';
 import {
   NewPostWrapper,
   Input,
@@ -15,91 +18,32 @@ import {
   FileNameInput,
 } from './styles';
 
-const readFile = file => {
-  return new Promise(resolve => {
-    const reader = new FileReader();
-    reader.addEventListener('load', () => resolve(reader.result), false);
-    reader.readAsDataURL(file);
-  });
-};
-
-const blobToFile = (theBlob, fileName) => {
-  theBlob.lastModifiedDate = new Date();
-  theBlob.name = fileName;
-  return theBlob;
-};
-
-const newBlob = async dataURL => {
-  const croppedImageResponse = await fetch(`${dataURL}`);
-  const blob = await croppedImageResponse.blob();
-  return blob;
-};
-
-const isFileTypeImage = filename => {
-  const extensionRegex = /(.jpg|.gif|.jpeg|.png)$/i;
-  return extensionRegex.test(filename);
-};
-
-const minZoom = 1;
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case 'INPUT_IMAGE':
-      return {
-        ...state,
-        originalImage: action.value,
-      };
-    case 'INPUT_IMAGE_URL':
-      return {
-        ...state,
-        originalImageUrl: action.value,
-      };
-    case 'INPUT_CROPPED_AREA':
-      return {
-        ...state,
-        croppedAreaPixels: action.value,
-      };
-    case 'INPUT_CONTENT':
-      return {
-        ...state,
-        contentValue: action.value,
-      };
-    case 'CHANGE_ZOOM':
-      return {
-        ...state,
-        zoom: action.value,
-      };
-    case 'CHANGE_CROP':
-      return {
-        ...state,
-        crop: action.value,
-      };
-    default:
-      return null;
-  }
-};
+const MIN_ZOOM = 1;
+const MAX_ZOOM = 3;
+const ZOOM_STEP = 0.1;
+const CROP_SIZE = 615;
 
 const NewPostPage = () => {
   const initialState = {
     originalImage: '',
     originalImageUrl: null,
     crop: { x: 0, y: 0 },
-    zoom: minZoom,
+    zoom: MIN_ZOOM,
     croppedAreaPixels: null,
     contentValue: '',
   };
-  const [state, dispatch] = useReducer(reducer, initialState);
+
+  const [
+    state,
+    inputImage,
+    onCropComplete,
+    changeContent,
+    changeCrop,
+    changeZoom,
+    changeSliderZoom,
+  ] = useImage(initialState);
   const [isSuccess, setSuccess] = useState(false);
   const [loading, setLoading] = useState(false);
-
-  const inputImage = async e => {
-    if (e.target.files && e.target.files.length > 0) {
-      dispatch({ type: 'INPUT_IMAGE', value: e.target.files[0] });
-
-      const originalImageDataUrl = await readFile(e.target.files[0]);
-      dispatch({ type: 'INPUT_IMAGE_URL', value: originalImageDataUrl });
-    }
-  };
 
   const post = useCallback(async () => {
     if (loading) return;
@@ -116,21 +60,12 @@ const NewPostPage = () => {
 
     try {
       setLoading(true);
-      const croppedImageDataUrl = await getCroppedImg(
-        state.originalImageUrl,
-        state.croppedAreaPixels,
-      );
-      const blob = await newBlob(croppedImageDataUrl);
-      const croppedImageBolb = blobToFile(blob, `${state.originalImage.name}`);
-      const croppedImageFile = new File(
-        [croppedImageBolb],
-        `${state.originalImage.name}`,
-        { type: 'image/png' },
-      );
+      const croppedImageFile = await makeNewImageFile(state);
       const formData = new FormData();
       formData.append('file', croppedImageFile);
       formData.append('content', state.contentValue);
       formData.append('userId', 1);
+
       const resultJSON = await fetch(
         `${process.env.REACT_APP_API_URL}/upload`,
         {
@@ -142,73 +77,46 @@ const NewPostPage = () => {
       if (result.result === 'success') {
         setSuccess(true);
       }
-    } catch (e) {}
+    } catch (e) {
+      return <Error />;
+    }
   }, [loading, state]);
-
-  const onCropComplete = useCallback(
-    (croppedArea, currentcroppedAreaPixels) => {
-      dispatch({ type: 'INPUT_CROPPED_AREA', value: currentcroppedAreaPixels });
-    },
-    [],
-  );
-
-  const changeContent = e => {
-    dispatch({ type: 'INPUT_CONTENT', value: e.target.value });
-  };
 
   if (isSuccess) {
     return <Redirect to="/" />;
   }
-
   return (
     <NewPostWrapper>
       {loading && <Loading size={50} />}
       <div className="section">
-        <FileNameInput
-          type="text"
-          value={state.originalImage.name}
-          disabled="disabled"
-        />
+        <FileNameInput value={state.originalImage.name} />
         <FileSelectLabel htmlFor="select_file">파일선택</FileSelectLabel>
         <FileInput onChange={inputImage} />
       </div>
       {state.originalImage && (
         <>
-          <div
-            className="crop-container"
-            style={{
-              position: 'relative',
-              width: '650px',
-              height: '650px',
-            }}
-          >
+          <div className="crop-container">
             <Cropper
-              minZoom={minZoom}
+              minZoom={MIN_ZOOM}
               image={state.originalImageUrl}
               crop={state.crop}
               zoom={state.zoom}
               aspect={1 / 1}
               restrictPosition={false}
-              onCropChange={crop =>
-                dispatch({ type: 'CHANGE_CROP', value: crop })
-              }
+              onCropChange={changeCrop}
               onCropComplete={onCropComplete}
-              onZoomChange={zoom =>
-                dispatch({ type: 'CHANGE_ZOOM', value: zoom })
-              }
-              cropSize={{ width: 615, height: 615 }}
+              onZoomChange={changeZoom}
+              cropSize={{ width: CROP_SIZE, height: CROP_SIZE }}
             />
           </div>
-          <div className="controls" style={{ width: '20%' }}>
+          <div className="controls">
             <Slider
               value={state.zoom}
-              min={minZoom}
-              max={3}
-              step={0.1}
+              min={MIN_ZOOM}
+              max={MAX_ZOOM}
+              step={ZOOM_STEP}
               aria-labelledby="Zoom"
-              onChange={(e, currentzoom) =>
-                dispatch({ type: 'CHANGE_ZOOM', value: currentzoom })
-              }
+              onChange={changeSliderZoom}
             />
           </div>
         </>
