+++ web/src/containers/NewPostPage/lib/cropImage.js
